condense files back
